+++ src/_incydr_sdk/enums/cases.py
@@ -2,6 +2,8 @@ from _incydr_sdk.enums import _Enum
 
 
 class SortKeys(_Enum):
+    """Possible keys to sort cases list results by."""
+
     NAME = "name"
     NUMBER = "number"
     CREATED_AT = "createdAt"
@@ -12,12 +14,16 @@ class SortKeys(_Enum):
 
 
 class CaseStatus(_Enum):
+    """Possible statuses for a case."""
+
     CLOSED = "CLOSED"
     OPEN = "OPEN"
     ARCHIVED = "ARCHIVED"
 
 
 class FileAvailability(_Enum):
+    """Possible states of file-event source file availability."""
+
     error = "ERROR"
     exact_file_available = "EXACT_FILE_AVAILABLE"
     no_file_available = "NO_FILE_AVAILABLE"